Include client details in provider contract
Oauth client for external providers should not exist. Include client information in oauth-provider contract.

Change-type: major
