+++ package.json
@@ -48,14 +48,14 @@
   "dependencies": {
     "@balena/jellyfish-assert": "^1.2.29",
     "@balena/jellyfish-environment": "^12.0.0",
-    "@balena/jellyfish-worker": "^24.2.2",
+    "@balena/jellyfish-worker": "^26.0.0",
     "is-uuid": "^1.0.2",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
     "@balena/jellyfish-config": "^2.0.6",
-    "@balena/jellyfish-plugin-default": "^25.1.0",
-    "@balena/jellyfish-plugin-product-os": "^5.0.24",
+    "@balena/jellyfish-plugin-default": "^27.0.0",
+    "@balena/jellyfish-plugin-product-os": "^7.0.0",
     "@balena/jellyfish-types": "^2.0.5",
     "@balena/lint": "^6.2.0",
     "@types/is-uuid": "^1.0.0",
